# sha-256 perf