@pdimov 2024-12-30
+17% isn't bad at all.